chore: renames 'FilterByCollaborationsStatus' to 'FilterByAcademyAndStatus'
