+++ server/src/collaborations/collaborations.module.ts
@@ -3,7 +3,7 @@ import { ICollaborationsRepository } from '@collaborations/interfaces';
 import { Collaboration } from '@collaborations/infra/typeorm/entities/collaboration.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Module } from '@nestjs/common';
-import { FilterByAcademyAndStatus } from './use-cases/filterByAcademyAndStatus/filter-collaborations-by-status';
+import { FilterByAcademyAndStatus } from './use-cases/filterByAcademyAndStatus/filter-by-academy-and-status';
 import { CollaborationsController } from 'shared/infra/http/controllers/collaborations.controller';
 
 @Module({
