+++ server/src/collaborations/dto/filter-by-academy-and-status.ts
@@ -1,6 +1,6 @@
 import { IsEnum, IsNotEmpty, IsString } from 'class-validator';
-import { CollaborationsStatus } from 'shared/constants';
-export class FilterCollaborationByStatusDTO {
+import { CollaborationsStatus } from '@shared/constants';
+export class FilterByAcademyAndStatusDTO {
   @IsNotEmpty()
   @IsEnum(CollaborationsStatus)
   status: CollaborationsStatus;
