+++ server/src/collaborations/dto/update-collaboration.dto.ts
@@ -1,6 +1,6 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateCollaborationDto } from './create-collaboration.dto';
+import { CreateCollaborationDTO } from './create-collaboration.dto';
 
-export class UpdateCollaborationDto extends PartialType(
-  CreateCollaborationDto,
+export class UpdateCollaborationDTO extends PartialType(
+  CreateCollaborationDTO,
 ) {}
