+++ server/src/shared/infra/http/controllers/collaborations.controller.ts
@@ -1,18 +1,16 @@
 import { Body, Controller, Get } from '@nestjs/common';
-import { FilterCollaborationByStatusDTO } from '@collaborations/dto/filter-collaboration-by-status.dto';
-import { FilterCollaborationsByStatus } from '@collaborations/use-cases/filterCollaborationsByStatus/filter-collaborations-by-status';
+import { FilterByAcademyAndStatusDTO } from '@collaborations/dto/filter-by-academy-and-status';
+import { FilterByAcademyAndStatus } from '@collaborations/use-cases/filterByAcademyAndStatus/filter-by-academy-and-status';
 
 @Controller('collaborations')
 export class CollaborationsController {
-  constructor(
-    private filterCollaborationsByStatus: FilterCollaborationsByStatus,
-  ) {}
+  constructor(private filterByAcademyAndStatus: FilterByAcademyAndStatus) {}
 
   @Get('academy/:id')
   filterByStatus(
     @Body()
-    { status, academy_id }: FilterCollaborationByStatusDTO,
+    { status, academy_id }: FilterByAcademyAndStatusDTO,
   ) {
-    return this.filterCollaborationsByStatus.execute({ status, academy_id });
+    return this.filterByAcademyAndStatus.execute({ status, academy_id });
   }
 }
